+++ CookBook.Recipes/database/dbo/Tables/RecipeInstructions.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[RecipeInstructions]
 (
-    [Id]                TINYINT IDENTITY(1, 1) NOT NULL,
+    [Id]                BIGINT IDENTITY(1, 1) NOT NULL,
     [RecipeId]          BIGINT NOT NULL,
     [Note]              NVARCHAR(1024) NOT NULL,
     [OrderIndex]        TINYINT NOT NULL DEFAULT 0,
